backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
websockets==12.0
sqlalchemy==2.0.27 
python-jose==3.3.0
passlib==1.7.4
python-multipart
bcrypt